refactor startup tasks and title mgmt related fucntions in StartupUtils and TitleUtils namespaces
