+++ include/icon.h
@@ -1,6 +1,8 @@
-unsigned int icon_tga_len = 11326;
+#pragma once
 
-unsigned char icon_tga[] = {
+inline unsigned int icon_tga_len = 11326;
+
+inline unsigned char icon_tga[] = {
 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC8, 0x00, 0x4B, 0x00, 
 0x20, 0x08, 0xA1, 0xD8, 0xD8, 0xD8, 0xFF, 0x02, 0xDE, 0xDB, 0xAA, 0xFF, 0xE0, 0xDC, 0x66, 0xFF, 
 0xE5, 0xE2, 0x8F, 0xFF, 0x83, 0xD8, 0xD8, 0xD8, 0xFF, 0x01, 0xDE, 0xD9, 0x9E, 0xFF, 0xD9, 0xD9,
