+++ include/utils/ConsoleUtils.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#define __STDC_WANT_LIB_EXT2__ 1
+
 #include <string>
 #include <utils/Colors.h>
 #include <utils/DrawUtils.h>
