+++ include/utils/StartupUtils.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace StartupUtils {
+
+    void getWiiUSerialId();
+
+    void disclaimer();
+
+    void addInitMessage(const char *newMessage);
+    void resetMessageList();
+    void addInitMessageWithIcon(const char *newMessage);
+
+
+} // namespace StartupUtils
