+++ src/menu/TitleListState.cpp
@@ -80,12 +80,12 @@ ApplicationState::eSubState TitleListState::update(Input *input) {
             return SUBSTATE_RETURN;
         if (input->get(ButtonState::TRIGGER, Button::R)) {
             this->titleSort = (this->titleSort + 1) % 4;
-            sortTitle(this->titles, this->titles + this->titlesCount, this->titleSort, this->sortAscending);
+            TitleUtils::sortTitle(this->titles, this->titles + this->titlesCount, this->titleSort, this->sortAscending);
         }
         if (input->get(ButtonState::TRIGGER, Button::L)) {
             if (this->titleSort > 0) {
                 this->sortAscending = !this->sortAscending;
-                sortTitle(this->titles, this->titles + this->titlesCount, this->titleSort, this->sortAscending);
+                TitleUtils::sortTitle(this->titles, this->titles + this->titlesCount, this->titleSort, this->sortAscending);
             }
         }
         if (input->get(ButtonState::TRIGGER, Button::A)) {
